+++ packages/terra-alert/src/terra-dev-site/test/alert/DefaultAlert.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Alert from '../../../Alert';
-
-export default () => <Alert id="defaultAlert">Default Alert</Alert>;