+++ packages/terra-core-docs/src/terra-dev-site/test/alert/DefaultAlert.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Alert from 'terra-alert';
+
+export default () => <Alert id="defaultAlert">Default Alert</Alert>;